Add files via upload
This project automatically detects the smell of gas in the room and emits a warning signal. The alarm sounds simultaneously and a message is sent to the homeowner’s phone. In both cases, if the homeowner does not hear the alarm, the appliance automatically shuts off the gas from the mains and it can clean the air of room! 

By the way, this gas detector device can be manufactured in two directions for homes and for gas-powered cars.